backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
python-multipart
anthropic
gtts
aiofiles
python-dotenv
pydantic
httpx
numpy
pillow
openai-whisper
spacy
opencv-python-headless
mediapipe
boto3
